backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Eladanius"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.13"
fastapi = {extras = ["standard"], version = "^0.115.4"}
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
authlib = "^1.3.2"
uvicorn = "^0.32.0"
itsdangerous = "^2.2.0"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
